Renamed parser->codeparser to avoid absurd C related OOM condition when compiling on linux. Also added pytest.ini
